SMTP authentication backend.
see #989
