+++ modoboa/lib/authbackends.py
@@ -1,29 +1,64 @@
 """Custom authentication backends."""
 
-from django.contrib.auth.backends import ModelBackend
+import smtplib
+
+from django.conf import settings
 
 from modoboa.core.models import User
 from modoboa.lib.email_utils import split_mailbox
 from modoboa.parameters import tools as param_tools
 
 
-class SimpleBackend(ModelBackend):
-    """Simple authentication backend."""
+class SMTPBackend(object):
+    """A backend to authenticate against an SMTP server."""
 
     def authenticate(self, username=None, password=None):
+        """Check the username/password and return a User."""
+        if type(username) is unicode:
+            username = username.encode("utf-8")
+        if type(password) is unicode:
+            password = password.encode("utf-8")
+
+        host = getattr(settings, "AUTH_SMTP_SERVER_ADDRESS", "localhost")
+        port = getattr(settings, "AUTH_SMTP_SERVER_PORT", 25)
+        secured_mode = getattr(settings, "AUTH_SMTP_SECURED_MODE", None)
+        if secured_mode == "ssl":
+            smtp = smtplib.SMTP_SSL(host, port)
+        else:
+            smtp = smtplib.SMTP(host, port)
+            if secured_mode == "starttls":
+                smtp.starttls()
         try:
-            user = User.objects.get(username=username)
-        except User.DoesNotExist:
+            smtp.login(username, password)
+        except smtplib.SMTPException:
             return None
-        if not user.check_password(password):
-            return None
+        return self.get_or_create_user(username)
+
+    def get_or_create_user(self, username):
+        """Get a user or create it the first time.
+
+        .. note::
+
+           We assume the username is a valid email address.
+        """
+        user, created = User.objects.get_or_create(
+            username__iexact=username, defaults={
+                "username": username.lower(), "email": username.lower()
+            }
+        )
+        if created:
+            populate_callback(user)
         return user
 
-    def get_user(self, user_id):
+    def get_user(self, user_pk):
+        """Retrieve a User instance."""
+        user = None
         try:
-            return User.objects.get(pk=user_id)
+            user = User.objects.get(pk=user_pk)
         except User.DoesNotExist:
-            return None
+            pass
+        return user
+
 
 try:
     from django_auth_ldap.backend import (
